+++ sdk/identity/identity-broker/src/index.ts
@@ -13,12 +13,18 @@ import { NativeBrokerPlugin } from "@azure/msal-node-extensions";
  *
  * Example:
  *
- * ```typescript
- * import { useIdentityPlugin, DeviceCodeCredential } from "@azure/identity";
+ * ```ts snippet:using_plugins
+ * import { useIdentityPlugin, InteractiveBrowserCredential } from "@azure/identity";
  * import { nativeBrokerPlugin } from "@azure/identity-broker";
  *
- * // Load the plugin
  * useIdentityPlugin(nativeBrokerPlugin);
+ * const credential = new InteractiveBrowserCredential({
+ *   brokerOptions: {
+ *     enabled: true,
+ *     parentWindowHandle: new Uint8Array(0), // This should be a handle to the parent window
+ *   },
+ * });
+ *
  * ```
  */