+++ sdk/identity/identity/src/credentials/deviceCodeCredential.ts
@@ -48,14 +48,17 @@ export class DeviceCodeCredential implements TokenCredential {
    *
    * Developers can configure how this message is shown by passing a custom `userPromptCallback`:
    *
-   * ```js
+   * ```ts snippet:device_code_credential_example
+   * import { DeviceCodeCredential } from "@azure/identity";
+   *
    * const credential = new DeviceCodeCredential({
-   *   tenantId: env.AZURE_TENANT_ID,
-   *   clientId: env.AZURE_CLIENT_ID,
+   *   tenantId: process.env.AZURE_TENANT_ID,
+   *   clientId: process.env.AZURE_CLIENT_ID,
    *   userPromptCallback: (info) => {
    *     console.log("CUSTOMIZED PROMPT CALLBACK", info.message);
-   *   }
+   *   },
    * });
+   *
    * ```
    *
    * @param options - Options for configuring the client which makes the authentication requests.