+++ sdk/notificationhubs/notification-hubs/eslint.config.mjs
@@ -8,4 +8,11 @@ export default [
       "n/no-process-exit": "off",
     },
   },
+  {
+    files: ["test/snippets.spec.ts"],
+    rules: {
+      "@typescript-eslint/ban-ts-comment": "off",
+      "@typescript-eslint/no-unused-vars": "off",
+    },
+  },
 ];