+++ sdk/ai/ai-inference-rest/samples/v1-beta/typescript/src/telemetryWithToolCall.ts
@@ -19,7 +19,9 @@ import * as dotenv from "dotenv";
 dotenv.config();
 
 // You will need to set these environment variables or edit the following values
-const modelEndpoint = process.env["MODEL_ENDPOINT"] || "<endpoint>";
+const endpoint = process.env["ENDPOINT"] || "<endpoint>";
+const key = process.env["KEY"];
+const modelName = process.env["MODEL_NAME"];
 const connectionString = process.env["APPLICATIONINSIGHTS_CONNECTION_STRING"];
 
 const provider = new NodeTracerProvider();
@@ -55,7 +57,7 @@ const getCurrentWeather = {
 };
 
 const getWeatherFunc = (location: string, unit: string): string => {
-  if (unit != "celsius") {
+  if (unit !== "celsius") {
     unit = "fahrenheit";
   }
   return `The temperature in ${location} is 72 degrees ${unit}`;
@@ -111,21 +113,17 @@ const handleToolCalls = (functionArray: Array<any>) => {
 
 // any import such as ai-inference has core-tracing as dependency must be imported after the instrumentation is registered
 import ModelClient, { ChatRequestMessage, isUnexpected } from "@azure-rest/ai-inference";
+import { AzureKeyCredential } from "@azure/core-auth";
 
 export async function main() {
-  const credential = new DefaultAzureCredential();
-  // auth scope for AOAI resources is currently https://cognitiveservices.azure.com/.default
-  // (only needed when targetting AOAI, do not use for Serverless API or Managed Computer Endpoints)
-  const scopes = ["https://cognitiveservices.azure.com/.default"];
-  const clientOptions = { credentials: { scopes } };
-  const client = ModelClient(modelEndpoint, credential, clientOptions);
+  const client = createModelClient();
 
   const messages: ChatRequestMessage[] = [{ role: "user", content: "What's the weather like in Boston?" }];
 
   let toolCallAnswer = "";
   let awaitingToolCallAnswer = true;
 
-  const tracer = trace.getTracer('sample', '0.1.0');
+  const tracer = trace.getTracer("sample", "0.1.0");
 
   await tracer.startActiveSpan('main', async (span) => {
     while (awaitingToolCallAnswer) {
@@ -138,7 +136,8 @@ export async function main() {
               type: "function",
               function: getCurrentWeather,
             },
-          ]
+          ],
+          model: modelName
         },
         tracingOptions: { tracingContext: context.active() }
       });
@@ -187,9 +186,22 @@ export async function main() {
 
   console.log("Model response after tool call:");
   console.log(toolCallAnswer);
-
 }
 
+/*
+  * This function creates a model client.
+  */
+function createModelClient() {
+  // auth scope for AOAI resources is currently https://cognitiveservices.azure.com/.default
+  // (only needed when targetting AOAI, do not use for Serverless API or Managed Computer Endpoints)
+  if (key) {
+    return ModelClient(endpoint, new AzureKeyCredential(key));      
+  } else {
+    const scopes = ["https://cognitiveservices.azure.com/.default"];
+    const clientOptions = { credentials: { scopes } };
+    return ModelClient(endpoint, new DefaultAzureCredential(), clientOptions);      
+  }
+}
 main().catch((err) => {
   console.error("The sample encountered an error:", err);
 });