+++ sdk/keyvault/keyvault-common/src/keyVaultAuthenticationPolicy.ts
@@ -2,15 +2,17 @@
 // Licensed under the MIT License.
 
 import {
-  AuthorizeRequestOnChallengeOptions,
-  AuthorizeRequestOptions,
-  ChallengeCallbacks,
+  PipelinePolicy,
   PipelineRequest,
+  PipelineResponse,
   RequestBodyType,
+  SendRequest,
 } from "@azure/core-rest-pipeline";
 import { WWWAuthenticate, parseWWWAuthenticateHeader } from "./parseWWWAuthenticate.js";
 
-import { GetTokenOptions } from "@azure/core-auth";
+import { GetTokenOptions, TokenCredential } from "@azure/core-auth";
+import { createTokenCycler } from "./tokenCycler.js";
+import { logger } from "./logger.js";
 
 /**
  * @internal
@@ -35,12 +37,13 @@ type ChallengeState =
   | {
       status: "complete";
       scopes: string[];
+      tenantId?: string;
     };
 
 /**
  * Additional options for the challenge based authentication policy.
  */
-export interface CreateChallengeCallbacksOptions {
+export interface KeyVaultAuthenticationPolicyOptions {
   /**
    * Whether to disable verification that the challenge resource matches the Key Vault or Managed HSM domain.
    *
@@ -67,7 +70,12 @@ function verifyChallengeResource(scope: string, request: PipelineRequest): void 
 }
 
 /**
- * Creates challenge callback handlers to manage CAE lifecycle in Azure Key Vault.
+ * Name of the Key Vault authentication policy.
+ */
+export const keyVaultAuthenticationPolicyName = "keyVaultAuthenticationPolicy";
+
+/**
+ * A custom implementation of the bearer-token authentication policy that handles Key Vault and CAE challenges.
  *
  * Key Vault supports other authentication schemes, but we ensure challenge authentication
  * is used by first sending a copy of the request, without authorization or content.
@@ -79,11 +87,13 @@ function verifyChallengeResource(scope: string, request: PipelineRequest): void 
  * if possible.
  *
  */
-export function createKeyVaultChallengeCallbacks(
-  options: CreateChallengeCallbacksOptions = {},
-): ChallengeCallbacks {
+export function keyVaultAuthenticationPolicy(
+  credential: TokenCredential,
+  options: KeyVaultAuthenticationPolicyOptions = {},
+): PipelinePolicy {
   const { disableChallengeResourceVerification } = options;
   let challengeState: ChallengeState = { status: "none" };
+  const getAccessToken = createTokenCycler(credential);
 
   function requestToOptions(request: PipelineRequest): GetTokenOptions {
     return {
@@ -95,10 +105,7 @@ export function createKeyVaultChallengeCallbacks(
     };
   }
 
-  async function authorizeRequest({
-    request,
-    getAccessToken,
-  }: AuthorizeRequestOptions): Promise<void> {
+  async function authorizeRequest(request: PipelineRequest): Promise<void> {
     const requestOptions: GetTokenOptions = requestToOptions(request);
 
     switch (challengeState.status) {
@@ -112,21 +119,29 @@ export function createKeyVaultChallengeCallbacks(
       case "started":
         break; // Retry, we should not overwrite the original body
       case "complete": {
-        const token = await getAccessToken(challengeState.scopes, requestOptions);
+        const token = await getAccessToken(challengeState.scopes, {
+          ...requestOptions,
+          enableCae: true,
+          tenantId: challengeState.tenantId,
+        });
         if (token) {
           request.headers.set("authorization", `Bearer ${token.token}`);
         }
         break;
       }
     }
-    return Promise.resolve();
   }
 
-  async function authorizeRequestOnChallenge({
-    request,
-    response,
-    getAccessToken,
-  }: AuthorizeRequestOnChallengeOptions): Promise<boolean> {
+  async function handleChallenge(
+    request: PipelineRequest,
+    response: PipelineResponse,
+    next: SendRequest,
+  ): Promise<PipelineResponse> {
+    // If status is not 401, this is a no-op
+    if (response.status !== 401) {
+      return response;
+    }
+
     if (request.body === null && challengeState.status === "started") {
       // Reset the original body before doing anything else.
       // Note: If successful status will be "complete", otherwise "none" will
@@ -138,16 +153,20 @@ export function createKeyVaultChallengeCallbacks(
 
     const challenge = response.headers.get("WWW-Authenticate");
     if (!challenge) {
-      throw new Error("Missing challenge.");
+      logger.warning(
+        "keyVaultAuthentication policy encountered a 401 response without a corresponding WWW-Authenticate header. This is unexpected. Not handling the 401 response.",
+      );
+      return response;
     }
-    const parsedChallenge: WWWAuthenticate = parseWWWAuthenticateHeader(challenge) || {};
+    const parsedChallenge: WWWAuthenticate = parseWWWAuthenticateHeader(challenge);
 
     const scope = parsedChallenge.resource
       ? parsedChallenge.resource + "/.default"
       : parsedChallenge.scope;
 
     if (!scope) {
-      throw new Error("Missing scope.");
+      // Cannot handle this kind of challenge here (if scope is not present, may be a CAE challenge)
+      return response;
     }
 
     if (!disableChallengeResourceVerification) {
@@ -156,11 +175,13 @@ export function createKeyVaultChallengeCallbacks(
 
     const accessToken = await getAccessToken([scope], {
       ...getTokenOptions,
+      enableCae: true,
       tenantId: parsedChallenge.tenantId,
     });
 
     if (!accessToken) {
-      return false;
+      // No access token provided, treat as no-op
+      return response;
     }
 
     request.headers.set("Authorization", `Bearer ${accessToken.token}`);
@@ -168,13 +189,76 @@ export function createKeyVaultChallengeCallbacks(
     challengeState = {
       status: "complete",
       scopes: [scope],
+      tenantId: parsedChallenge.tenantId,
     };
 
-    return true;
+    // We have a token now, so try send the request again
+    return next(request);
+  }
+
+  async function handleCaeChallenge(
+    request: PipelineRequest,
+    response: PipelineResponse,
+    next: SendRequest,
+  ): Promise<PipelineResponse> {
+    // Cannot handle CAE challenge if a regular challenge has not been completed first
+    if (challengeState.status !== "complete") {
+      return response;
+    }
+
+    // If status is not 401, this is a no-op
+    if (response.status !== 401) {
+      return response;
+    }
+
+    const getTokenOptions = requestToOptions(request);
+
+    const challenge = response.headers.get("WWW-Authenticate");
+    if (!challenge) {
+      return response;
+    }
+    const { claims: base64EncodedClaims, error }: WWWAuthenticate =
+      parseWWWAuthenticateHeader(challenge);
+
+    if (error !== "insufficient_claims" || base64EncodedClaims === undefined) {
+      return response;
+    }
+
+    const claims = atob(base64EncodedClaims);
+
+    const accessToken = await getAccessToken(challengeState.scopes, {
+      ...getTokenOptions,
+      enableCae: true,
+      tenantId: challengeState.tenantId,
+      claims,
+    });
+
+    request.headers.set("Authorization", `Bearer ${accessToken.token}`);
+
+    return next(request);
+  }
+
+  async function sendRequest(
+    request: PipelineRequest,
+    next: SendRequest,
+  ): Promise<PipelineResponse> {
+    // Add token if possible
+    await authorizeRequest(request);
+
+    // Try send request (first attempt)
+    let response = await next(request);
+
+    // Handle standard challenge if present
+    response = await handleChallenge(request, response, next);
+
+    // Handle CAE challenge if present
+    response = await handleCaeChallenge(request, response, next);
+
+    return response;
   }
 
   return {
-    authorizeRequest,
-    authorizeRequestOnChallenge,
+    name: keyVaultAuthenticationPolicyName,
+    sendRequest,
   };
 }