+++ sdk/identity/identity/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Release History
 
+<!-- dev-tool snippets ignore -->
 ## 4.5.0-beta.3 (2024-09-18)
 
 ### Features Added